Fix arithmetics for opacity. Count in 16-bit rather then 8-bit as integer overflow occurs with visual artefacts.
svn path=/trunk/koffice/; revision=1093726

+++ krita/plugins/paintops/particle/particle_brush.cpp
@@ -68,19 +68,19 @@ void ParticleBrush::paintParticle(KisRandomAccessor& accWrite,KoColorSpace * cs,
     quint8 bbr = qRound((fx)  * (fy)  * opacity * weight);
 
     accWrite.moveTo(ipx  , ipy);
-    myColor.setOpacity( qBound<quint8>(OPACITY_TRANSPARENT_U8,btl + cs->opacityU8(accWrite.rawData()),OPACITY_OPAQUE_U8) );
+    myColor.setOpacity( quint8(qBound<quint16>(OPACITY_TRANSPARENT_U8,btl + cs->opacityU8(accWrite.rawData()),OPACITY_OPAQUE_U8)) );
     memcpy(accWrite.rawData(), myColor.data(), cs->pixelSize());
 
     accWrite.moveTo(ipx + 1, ipy);
-    myColor.setOpacity( qBound<quint8>(OPACITY_TRANSPARENT_U8,btr + cs->opacityU8(accWrite.rawData()),OPACITY_OPAQUE_U8) );
+    myColor.setOpacity( quint8(qBound<quint16>(OPACITY_TRANSPARENT_U8,btr + cs->opacityU8(accWrite.rawData()),OPACITY_OPAQUE_U8)) );
     memcpy(accWrite.rawData(), myColor.data(), cs->pixelSize());
 
     accWrite.moveTo(ipx, ipy + 1);
-    myColor.setOpacity( qBound<quint8>(OPACITY_TRANSPARENT_U8,bbl + cs->opacityU8(accWrite.rawData()),OPACITY_OPAQUE_U8) );
+    myColor.setOpacity( quint8(qBound<quint16>(OPACITY_TRANSPARENT_U8,bbl + cs->opacityU8(accWrite.rawData()),OPACITY_OPAQUE_U8)) );
     memcpy(accWrite.rawData(), myColor.data(), cs->pixelSize());
 
     accWrite.moveTo(ipx + 1, ipy + 1);
-    myColor.setOpacity( qBound<quint8>(OPACITY_TRANSPARENT_U8,bbr + cs->opacityU8(accWrite.rawData()),OPACITY_OPAQUE_U8) );
+    myColor.setOpacity( quint8(qBound<quint16>(OPACITY_TRANSPARENT_U8,bbr + cs->opacityU8(accWrite.rawData()),OPACITY_OPAQUE_U8)) );
     memcpy(accWrite.rawData(), myColor.data(), cs->pixelSize());
 }
